+++ packages/plugin-packages/mashroom-websocket/src/backend/webapp/mashroom-bootstrap-webapp.js
@@ -13,17 +13,20 @@ import type {
     MashroomHttpUpgradeHandler,
     MashroomWebAppPluginBootstrapFunction
 } from '@mashroom/mashroom/type-definitions';
+import ReconnectMessageBufferStore from './ReconnectMessageBufferStore';
 
 const bootstrap: MashroomWebAppPluginBootstrapFunction = async (pluginName, pluginConfig, pluginContextHolder) => {
-    const { path, restrictToRoles, enableKeepAlive, keepAliveIntervalSec, maxConnections } = pluginConfig;
+    const { path, restrictToRoles, enableKeepAlive, keepAliveIntervalSec, maxConnections, reconnectMessageBufferFolder, reconnectTimeoutSec } = pluginConfig;
     const pluginContext = pluginContextHolder.getPluginContext();
 
     context.restrictToRoles = restrictToRoles;
     context.basePath = path;
     context.enableKeepAlive = enableKeepAlive;
     context.keepAliveIntervalSec = keepAliveIntervalSec;
+    context.reconnectTimeoutSec = reconnectTimeoutSec;
     context.maxConnections = maxConnections;
-    context.server = new WebSocketServer(pluginContext.loggerFactory);
+    const reconnectMessageBufferStore = new ReconnectMessageBufferStore(reconnectMessageBufferFolder, pluginContext.loggerFactory);
+    context.server = new WebSocketServer(pluginContext.loggerFactory, reconnectMessageBufferStore);
 
     const upgradeHandler: MashroomHttpUpgradeHandler = httpUpgradeHandlerFn();