+++ packages/plugin-packages/mashroom-websocket/src/frontend/test_client.js
@@ -19,7 +19,8 @@ global.connect = () => {
         const host = global.document.location.hostname;
         const port = global.document.location.port ? `:${global.document.location.port}` : '';
         const path = global.document.location.pathname;
-        const wsUrl = `${socketProtocol}//${host}${port}${path}`;
+        const clientId = global.sessionStorage.getItem('MashroomPortalWebSocketClientId');
+        const wsUrl = `${socketProtocol}//${host}${port}${path}${clientId ? `?clientId=${clientId}` : ''}`;
         console.info('Try to connect to:', wsUrl);
         webSocket = new global.WebSocket(wsUrl);
         webSocket.onopen = () => {
@@ -39,7 +40,22 @@ global.connect = () => {
             console.info('Received message', event);
             addLog(`Received message: ${String(event.data)}`);
             setStatus('Connected');
+            try {
+                const dataObj = JSON.parse(String(event.data));
+                const { type, payload } = dataObj;
+                if (type === 'setClientId') {
+                    global.sessionStorage.setItem('MashroomPortalWebSocketClientId', payload);
+                }
+            } catch (e) {
+                console.error('Cannot parse event payload:', e.message);
+            }
         };
+    }
+};
+
+global.disconnect = () => {
+    if (webSocket) {
+        webSocket.close();
     }
 };