+++ packages/plugin-packages/mashroom-websocket/type-definitions/index.d.ts
@@ -19,7 +19,9 @@ export type MashroomWebSocketClient = {
     readonly connectPath: string;
     readonly user: MashroomSecurityUser;
     readonly loggerContext: {};
+    readonly clientId: string;
     alive: boolean;
+    reconnecting?: TimeoutID;
 };
 
 /**