+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  * Improved reliability and performance of the LDAP security provider
  * Upgraded libraries with known vulnerabilities
  * Portal: [ReasonReact](https://reasonml.github.io/reason-react) based demo app added
+ * Added reconnect to previous session functionality for websockets
 
 ## 1.5.1 (August 25, 2020)